+++ activejob/lib/active_job/queue_adapters/que_adapter.rb
@@ -17,11 +17,15 @@ module ActiveJob
     class QueAdapter
       class << self
         def enqueue(job) #:nodoc:
-          JobWrapper.enqueue job.serialize, queue: job.queue_name
+          enqueue_at(job, nil)
         end
 
         def enqueue_at(job, timestamp) #:nodoc:
-          JobWrapper.enqueue job.serialize, queue: job.queue_name, run_at: Time.at(timestamp)
+          options = {}
+          options[:queue]    = job.queue_name
+          options[:priority] = job.priority if job.priority
+          options[:run_at]   = Time.at(timestamp) if timestamp
+          JobWrapper.enqueue job.serialize, options
         end
       end