+++ activejob/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Add job priorities to Active Job.
+
+    *wvengen*
+
 *   Allow keyword arguments to be used with Active Job.
 
     Fixes #18741.